рефакторинг: вынес обработчики в handlers, функции в services
