--- library/projects/teleban/text_content.py
@@ -1,20 +1,0 @@
-INFO_MSG = """
-Здравствуй, добрый человек!
-Я Бот Нужных Новостей!
-Я буду искать тебе только самые актуальные и интересные новости, чтоб ты не страдал от инфошума, но оставался в курсе событий.
-У тебя мало времени, Я в тебе это ценю, выбери нужные подписки и поехали!
-"""
-
-HELP_MSG = """
-🤖 <b>Справка по командам бота</b>
-        /start - Начать работу с ботом и открыть главную клавиатуру.
-        /help - Получить помощь и описание возможностей бота.
-        /subscriptions - Мои подписки.
-        /info - Информация о боте.
-"""
-
-SUBSCRIPTIONS_MSG = """
-Привет, {user_name}
-
-Здесь перечислены твои подписки:
-"""
