+++ library/projects/teleban/text_data.py
@@ -1,0 +1,31 @@
+START_MSG = """
+Привет, {user_name} 👋
+
+Я ТелеБАН - 🤖 бот нужных новостей!
+
+Я хочу избавить тебя от инфошума, поэтому буду присылать тебе только новости, которые ты выберешь <u>сам</u>.
+
+Для этого подпишись на категории новостей ниже в меню:
+"""
+
+INFO_MSG = """
+Вы можете подписаться на нужные категории новостей с популярных сайтов, и бот будет оперативно вам их присылать.
+
+Посмотреть текущие подписки можно в меню "Мои подписки".
+
+Для отправки замечаний, пожеланий и предложений просто напишите сообщение боту.
+"""
+
+HELP_MSG = """
+Справка по командам бота 💡
+
+/start - начать работу с ботом
+/help  - получить помощь
+/info - информация о боте
+/subs - посмотреть свои подписки
+"""
+
+SUBS_MSG = """
+{user_name},
+ниже перечислены твои подписки:
+"""
